+++ presentation/src/main/java/bisq/presentation/notifications/other/AwtNotificationSender.java
@@ -15,8 +15,9 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.presentation.notifications;
+package bisq.presentation.notifications.other;
 
+import bisq.presentation.notifications.NotificationSender;
 import lombok.extern.slf4j.Slf4j;
 
 import java.awt.*;
